+++ src/main/java/g1001_1100/s1009_complement_of_base_10_integer/Solution.java
@@ -20,7 +20,7 @@ public class Solution {
         int exp = list.size() - 1;
         for (int i = list.size() - 1; i >= 0; i--) {
             if (list.get(i) == 0) {
-                result += Math.pow(2, exp);
+                result += (int) Math.pow(2, exp);
             }
             exp--;
         }
