+++ src/main/java/g2501_2600/s2528_maximize_the_minimum_powered_city/Solution.java
@@ -19,7 +19,7 @@ public class Solution {
             if (req > k) {
                 return false;
             }
-            k -= req;
+            k -= (int) req;
             extraPower[i] += (req);
             if (i + 2 * r + 1 < n) {
                 extraPower[i + 2 * r + 1] -= (req);
